requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==3.2.7
django-cors-headers==4.1.0
django-jazzmin==2.6.0
django-username-email==2.5.4
geographiclib==2.0
geopy==2.4.0
googlemaps==4.10.0
gunicorn==21.2.0
idna==3.4
packaging==23.2
Pillow==9.5.0
python-dotenv==0.21.1
requests==2.31.0
sqlparse==0.4.4
stripe==7.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.5
whitenoise==6.5.0
